Add performance measurement task
With this change we introduce a separate Github action that periodically
runs a performance measurement and collects values after each run
via Theia's prometheus endpoint (`/metrics`).
These values are kept persistent over time in the github pages.

Also we generate an initial report with graphs that plot the playwright
execution time and the CPU seconds of each run.
This report will be available at:
https://eclipse-theia.github.io/theia-e2e-test-suite/performance

Contributed on behalf of STMicroelectronics