Publish allure reports into subdirectory
Update the CI workflow to publish allure reports into the `allure` subdirectory instead of the root.
This ensure that the cleanup behavior (remove old reports) of the corresponding github action does not interfere with other unrelated results. (e.g. removing the performance history directory instead of an old test directory)