Add simple script to convert the textual output to trace-event format
Add a script to scrape the github page and convert all the text logs to
the Trace Event Format readable by Chromium and Eclipse Trace Compass.

This allows the generation of flame graphs, flame charts and a comparison
of CPU usage and wall time. It also will show gaps in the instrumentation.
This in turn enables optimization efforts to be applied in the right areas.

Usage:

1- set up an environment (pip3 install -r requirements.txt)
2- run the script from a command line (python3 get_trace.py)
3- there will now be over 30 json files in the root folder either
 a- open them one by one in Chromium
 b- open them all at once in Eclipse Trace Compass
4- In the flame chart, look for dominators and apply optimizations there

Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>